requirements.txt
numpy>=1.23
scikit-learn>=1.3
Pillow>=9.0
